Reimplement and extract change detection
