+++ cli/package.json
@@ -31,10 +31,11 @@
     "findup-sync": "~0.4.3",
     "graphlib": "^2.1.1",
     "handlebars": "~4.0.6",
-    "ignore-file": "~1.1.2",
     "jsondiffpatch": "~0.2.4",
     "lodash": "~4.17.4",
+    "minimatch": "^3.0.3",
     "object-filter": "~1.0.2",
+    "parse-gitignore": "^0.3.1",
     "shelljs": "~0.7.5",
     "yargs": "~6.6.0"
   },
